manage: propertynotify: Reduce cost of unused size hint updates
This patch defers all size hint calculations until they are actually
needed, drastically reducing the number of calls to updatesizehints(),
which can be expensive when called repeatedly (as it currently is during
resizes).

In my unscientific testing this reduces calls to updatesizehints() by
over 90% during a typical work session. There are no functional changes
for users other than an increase in responsiveness after resizes and
reduction in CPU time.

There are two main motivations to defer this work to the time of hint
application:

1. Some clients, especially terminals using incremental size hints,
   resend XA_WM_NORMAL_HINTS on resize to avoid fighting with the WM or
   mouse resizing. For example, some terminals like urxvt clear
   PBaseSize and PResizeInc during XResizeWindow and restore them
   afterwards.

   For this reason we typically receive a backlogged XA_WM_NORMAL_HINTS
   message for each update period with movement, which is useless. In
   some cases one may get hundreds or thousands of XA_WM_NORMAL_HINTS
   messages on large resizes, and currently all of these result in a
   separate updatesizehints() call which is immediately invalidated.

   (We can't just blindly discard these messages during resizes like we
   do for EnterNotify, because some of them might actually be for other
   windows, or not be XA_WM_NORMAL_HINTS events.)

2. For users using resizehints=0 most of these updates are unused anyway
   -- in the normal case where the client is not floating these values
   won't be used, so there's no need to calculate them up front.

A synthetic test using the mouse to resize a floating terminal window
from roughly 256x256 to 1024x1024 and back again shows that the number
of calls to updatesizehints() goes from over 500 (one for each update
interval with movement) to 2 (one for each hint application), with no
change in user visible behaviour.

This also reduces the delay before dwm is ready to process new events
again after a large resize on such a client, as it avoids the thundering
herd of updatesizehints() calls when hundreds of backlogged
XA_WM_NORMAL_HINTS messages appear at once after a resize is finished.
In slower environments or X servers, this improvement in responsiveness
is often tangible after resizing a client that changes hints during
resizes.